Change PC selector form QLineEdit to QSpinBox
I think it is better to use a QspinBox instead of QlineEdit; it is easier to navigate using the spinbox. 
Also, for new users, it automatically directs them to the ability to change PCs.
I used/tested it in the last week, and it is working well on my computer (Linux).