comment sbxreader nchannels overwrite
I am not sure what is the correct way to design it. But I found that sometimes I don't want to analyze the red channel. 
If I set nchannels parameter to 1 in ops (and comment the channel overwrite here), I can analyze only the first channel. 

So I think that either the ops parameter should not be overwritten by sbxreader, or that when using sbxreader one can not give those parameters (It is actually very easy to extract them to the gui once a file is selected, and then gray-out those parameters).