fixing the conda create statement...
the conda create statement for the developer version was incorrect (it added the "env" argument unnecessarily) and didn't specify which python version to use. The newest python (3.11) has updates to the dataclasses module, which prevent the ROI class from working properly (in detection/stats.py) due to it's not being hashable.