CheckState fix for ROIs_On
state is returned as an integer (0 for unchecked, 2 for checked). When state is compared with `QtCore.Qt.Checked`, it always evaluates to False, because Qt.Checked is an enum, not an int. Processing state through QtCore.Qt.CheckState fixes this error and allows the user to toggle ROIs on and off